refactor: cc-0018 put the uat link into .env file
