+++ .gitignore
@@ -43,3 +43,7 @@ yarn-error.log*
 .history
 
 next-env.d.ts
+
+#.env 
+
+.env
